Follow the reworked API in the sample
The sample gains the translucent-activity and popup window kinds, reports
the active UnderlaySource and drives UnderlayState.refresh().

Tile bitmaps are drawn on Dispatchers.Default instead of inside
composition, where eight of them held up the first frame. UnderlayCard
and SampleScreen take a modifier, and the tile shape is hoisted out of
the per-tile composition.

+++ sample/src/main/kotlin/com/timkrest/underlay/sample/TranslucentOverlayActivity.kt
@@ -16,7 +16,7 @@ class TranslucentOverlayActivity : ComponentActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        val blurRadius = intent.getIntExtra(EXTRA_BLUR_RADIUS, DEFAULT_BLUR_RADIUS.value.toInt()).dp
+        val blurRadius = intent.getFloatExtra(EXTRA_BLUR_RADIUS, DEFAULT_BLUR_RADIUS.value).dp
 
         setContent {
             SampleTheme {
@@ -40,6 +40,6 @@ class TranslucentOverlayActivity : ComponentActivity() {
 
         fun intent(context: Context, blurRadius: Dp): Intent =
             Intent(context, TranslucentOverlayActivity::class.java)
-                .putExtra(EXTRA_BLUR_RADIUS, blurRadius.value.toInt())
+                .putExtra(EXTRA_BLUR_RADIUS, blurRadius.value)
     }
 }
